Reorganized and reworked all tests
