+++ openaerostruct/geometry/tests/multisection/test_joining_comp.py
@@ -3,12 +3,12 @@ import unittest
 
 from openaerostruct.geometry.geometry_multi_join import GeomMultiJoin
 from openaerostruct.geometry.geometry_group import build_sections
-from openaerostruct.utils.testing import run_test, get_three_section_surface
+from openaerostruct.utils.testing import run_test, get_three_section_surface_sym
 
 
 class Test(unittest.TestCase):
     def test(self):
-        (surface, chord_bspline) = get_three_section_surface()
+        (surface, chord_bspline) = get_three_section_surface_sym()
         sec_dicts = build_sections(surface)
 
         comp = GeomMultiJoin(sections=sec_dicts, dim_constr=[np.ones(3), np.ones(3), np.ones(3)])
