+++ openaerostruct/geometry/tests/multisection/test_unification_comp.py
@@ -2,12 +2,12 @@ import unittest
 
 from openaerostruct.geometry.geometry_unification import GeomMultiUnification
 from openaerostruct.geometry.geometry_group import build_sections
-from openaerostruct.utils.testing import run_test, get_three_section_surface
+from openaerostruct.utils.testing import run_test, get_three_section_surface_sym
 
 
 class Test(unittest.TestCase):
     def test(self):
-        (surface, chord_bspline) = get_three_section_surface()
+        (surface, chord_bspline) = get_three_section_surface_sym()
         sec_dicts = build_sections(surface)
 
         comp = GeomMultiUnification(sections=sec_dicts, surface_name=surface["name"])
