+++ openaerostruct/utils/testing.py
@@ -217,91 +217,7 @@ def get_ground_effect_surfaces():
     return surfaces
 
 
-def get_three_section_surfaces():
-    # Outputs a 3 section wing surface and a 2 section tail surface
-    # Set-up B-splines for each section. Done here since this information will be needed multiple times.
-    wing_chord_cp = [np.ones(2), np.ones(2), np.ones(2)]
-
-    tail_chord_cp = [np.ones(2), np.ones(2)]
-
-    wing_dict = {
-        # Wing definition
-        # Basic surface parameters
-        "name": "surface",
-        "isMultiSection": True,
-        "num_sections": 3,  # The number of sections in the multi-section surface
-        "sec_name": ["sec0", "sec1", "sec2"],  # names of the individual sections
-        "symmetry": True,  # if true, model one half of wing. reflected across the midspan of the root section
-        "S_ref_type": "wetted",  # how we compute the wing area,
-        # can be 'wetted' or 'projected'
-        # Geometry Parameters
-        "taper": [1.0, 1.0, 1.0],  # Wing taper for each section
-        "span": [1.0, 1.0, 1.0],  # Wing span for each section
-        "sweep": [0.0, 0.0, 0.0],  # Wing sweep for each section
-        "chord_cp": wing_chord_cp,  # Use previously set-up B-spline
-        "twist_cp": [np.zeros(2), np.zeros(2), np.zeros(2)],
-        "root_chord": 1.0,  # Wing root chord for each section
-        # Mesh Parameters
-        "meshes": "gen-meshes",  # Supply a mesh for each section or "gen-meshes" for automatic mesh generation
-        "nx": 3,  # Number of chordwise points. Same for all sections
-        "ny": [21, 21, 21],  # Number of spanwise points for each section
-        # Aerodynamic Parameters
-        "CL0": 0.0,  # CL of the surface at alpha=0
-        "CD0": 0.015,  # CD of the surface at alpha=0
-        # Airfoil properties for viscous drag calculation
-        "k_lam": 0.05,  # percentage of chord with laminar
-        # flow, used for viscous drag
-        "t_over_c_cp": [np.array([0.15]), np.array([0.15]), np.array([0.15])],  # thickness over chord ratio (NACA0015)
-        "c_max_t": 0.303,  # chordwise location of maximum (NACA0015)
-        # thickness
-        "with_viscous": True,  # if true, compute viscous drag
-        "with_wave": False,  # if true, compute wave drag
-        "groundplane": False,
-    }
-
-    tail_dict = {
-        # Wing definition
-        # Basic surface parameters
-        "name": "surface",
-        "isMultiSection": True,
-        "num_sections": 2,  # The number of sections in the multi-section surface
-        "sec_name": ["sec0", "sec1"],  # names of the individual sections
-        "symmetry": True,  # if true, model one half of wing. reflected across the midspan of the root section
-        "S_ref_type": "wetted",  # how we compute the wing area,
-        # can be 'wetted' or 'projected'
-        # Geometry Parameters
-        "taper": [1.0, 1.0],  # Wing taper for each section
-        "span": [1.0, 1.0],  # Wing span for each section
-        "sweep": [0.0, 0.0],  # Wing sweep for each section
-        "chord_cp": tail_chord_cp,  # Use previously set-up B-spline
-        "twist_cp": [np.zeros(2), np.zeros(2)],
-        # "sec_chord_cp": [np.ones(1),2*np.ones(1),3*np.ones(1)], #Chord B-spline control points for each section
-        "root_chord": 1.0,  # Wing root chord for each section
-        # Mesh Parameters
-        "meshes": "gen-meshes",  # Supply a mesh for each section or "gen-meshes" for automatic mesh generation
-        "nx": 3,  # Number of chordwise points. Same for all sections
-        "ny": [21, 21],  # Number of spanwise points for each section
-        # Aerodynamic Parameters
-        "CL0": 0.0,  # CL of the surface at alpha=0
-        "CD0": 0.015,  # CD of the surface at alpha=0
-        # Airfoil properties for viscous drag calculation
-        "k_lam": 0.05,  # percentage of chord with laminar
-        # flow, used for viscous drag
-        "t_over_c_cp": [np.array([0.15]), np.array([0.15])],  # thickness over chord ratio (NACA0015)
-        "c_max_t": 0.303,  # chordwise location of maximum (NACA0015)
-        # thickness
-        "with_viscous": True,  # if true, compute viscous drag
-        "with_wave": False,  # if true, compute wave drag
-        "groundplane": False,
-    }
-
-    surfaces = [wing_dict, tail_dict]
-    bsplines = [wing_chord_cp, tail_chord_cp]
-
-    return surfaces, bsplines
-
-
-def get_three_section_surface():
+def get_three_section_surface_sym():
     # Outputs a three section wing surface
     # Set-up B-splines for each section. Done here since this information will be needed multiple times.
     sec_chord_cp = [np.ones(2), np.ones(2), np.ones(2)]
@@ -319,24 +235,76 @@ def get_three_section_surface():
         # Geometry Parameters
         "taper": [1.0, 1.0, 1.0],  # Wing taper for each section
         "span": [1.0, 1.0, 1.0],  # Wing span for each section
-        "sweep": [0.0, 0.0, 0.0],  # Wing sweep for each section
-        "chord_cp": sec_chord_cp,  # Use previously set-up B-spline
+        "sweep": [0.0, 0, 0.0],  # Wing sweep for each section
         "twist_cp": [np.zeros(2), np.zeros(2), np.zeros(2)],
+        "chord_cp": sec_chord_cp,
         "root_chord": 1.0,  # Wing root chord for each section
         # Mesh Parameters
         "meshes": "gen-meshes",  # Supply a mesh for each section or "gen-meshes" for automatic mesh generation
-        "nx": 3,  # Number of chordwise points. Same for all sections
-        "ny": [21, 21, 21],  # Number of spanwise points for each section
+        "nx": 2,  # Number of chordwise points. Same for all sections
+        "ny": [11, 11, 11],  # Number of spanwise points for each section
         # Aerodynamic Parameters
         "CL0": 0.0,  # CL of the surface at alpha=0
         "CD0": 0.015,  # CD of the surface at alpha=0
         # Airfoil properties for viscous drag calculation
         "k_lam": 0.05,  # percentage of chord with laminar
         # flow, used for viscous drag
-        "t_over_c_cp": [np.array([0.15]), np.array([0.15]), np.array([0.15])],  # thickness over chord ratio (NACA0015)
+        "t_over_c_cp": [
+            np.array([0.15]),
+            np.array([0.15]),
+            np.array([0.15]),
+        ],  # thickness over chord ratio (NACA0015)
         "c_max_t": 0.303,  # chordwise location of maximum (NACA0015)
         # thickness
-        "with_viscous": True,  # if true, compute viscous drag
+        "with_viscous": False,  # if true, compute viscous drag
+        "with_wave": False,  # if true, compute wave drag
+        "groundplane": False,
+    }
+
+    return surface_dict, sec_chord_cp
+
+
+def get_three_section_surface_asym():
+    # Outputs a three section wing surface
+    # Set-up B-splines for each section. Done here since this information will be needed multiple times.
+    sec_chord_cp = [np.ones(2), np.ones(2), np.ones(2)]
+
+    surface_dict = {
+        # Wing definition
+        # Basic surface parameters
+        "name": "surface",
+        "isMultiSection": True,
+        "num_sections": 3,  # The number of sections in the multi-section surface
+        "sec_name": ["sec0", "sec1", "sec2"],  # names of the individual sections
+        "symmetry": False,  # if true, model one half of wing. reflected across the midspan of the root section
+        "S_ref_type": "wetted",  # how we compute the wing area,
+        # can be 'wetted' or 'projected'
+        "rootSection": 1,
+        # Geometry Parameters
+        "taper": [1.0, 1.0, 1.0],  # Wing taper for each section
+        "span": [1.0, 1.0, 1.0],  # Wing span for each section
+        "sweep": [0.0, 0.0, 0.0],  # Wing sweep for each section
+        "twist_cp": [np.zeros(2), np.zeros(2), np.zeros(2)],
+        "chord_cp": sec_chord_cp,
+        "root_chord": 1.0,  # Wing root chord for each section
+        # Mesh Parameters
+        "meshes": "gen-meshes",  # Supply a mesh for each section or "gen-meshes" for automatic mesh generation
+        "nx": 2,  # Number of chordwise points. Same for all sections
+        "ny": [11, 11, 11],  # Number of spanwise points for each section
+        # Aerodynamic Parameters
+        "CL0": 0.0,  # CL of the surface at alpha=0
+        "CD0": 0.015,  # CD of the surface at alpha=0
+        # Airfoil properties for viscous drag calculation
+        "k_lam": 0.05,  # percentage of chord with laminar
+        # flow, used for viscous drag
+        "t_over_c_cp": [
+            np.array([0.15]),
+            np.array([0.15]),
+            np.array([0.15]),
+        ],  # thickness over chord ratio (NACA0015)
+        "c_max_t": 0.303,  # chordwise location of maximum (NACA0015)
+        # thickness
+        "with_viscous": False,  # if true, compute viscous drag
         "with_wave": False,  # if true, compute wave drag
         "groundplane": False,
     }
@@ -390,6 +358,52 @@ def get_two_section_surface_sym():
     return surface, sec_chord_cp
 
 
+def get_two_section_surface_sym_visc():
+    # Outputs a symmetric two section wing surface
+    # Set-up B-splines for each section. Done here since this information will be needed multiple times.
+    sec_chord_cp = [np.array([1, 1]), np.array([1.0, 1.0])]
+
+    # Create a dictionary with info and options about the multi-section aerodynamic
+    # lifting surface
+    surface_dict = {
+        # Wing definition
+        # Basic surface parameters
+        "name": "surface",
+        "isMultiSection": True,
+        "num_sections": 2,  # The number of sections in the multi-section surface
+        "sec_name": ["sec0", "sec1"],  # names of the individual sections
+        "symmetry": True,  # if true, model one half of wing. reflected across the midspan of the root section
+        "S_ref_type": "wetted",  # how we compute the wing area,
+        # can be 'wetted' or 'projected'
+        # Geometry Parameters
+        "taper": [1.0, 1.0],  # Wing taper for each section
+        "span": [1.0, 1.0],  # Wing span for each section
+        "sweep": [0.0, 0.0],  # Wing sweep for each section
+        "chord_cp": sec_chord_cp,  # Use previously set-up B-spline
+        "twist_cp": [np.zeros(2), np.zeros(2)],
+        # "sec_chord_cp": [np.ones(1),2*np.ones(1),3*np.ones(1)], #Chord B-spline control points for each section
+        "root_chord": 1.0,  # Wing root chord for each section
+        # Mesh Parameters
+        "meshes": "gen-meshes",  # Supply a mesh for each section or "gen-meshes" for automatic mesh generation
+        "nx": 3,  # Number of chordwise points. Same for all sections
+        "ny": [21, 21],  # Number of spanwise points for each section
+        # Aerodynamic Parameters
+        "CL0": 0.0,  # CL of the surface at alpha=0
+        "CD0": 0.015,  # CD of the surface at alpha=0
+        # Airfoil properties for viscous drag calculation
+        "k_lam": 0.05,  # percentage of chord with laminar
+        # flow, used for viscous drag
+        "t_over_c_cp": [np.array([0.15]), np.array([0.15])],  # thickness over chord ratio (NACA0015)
+        "c_max_t": 0.303,  # chordwise location of maximum (NACA0015)
+        # thickness
+        "with_viscous": True,  # if true, compute viscous drag
+        "with_wave": False,  # if true, compute wave drag
+        "groundplane": False,
+    }
+
+    return surface_dict, sec_chord_cp
+
+
 def get_two_section_surface_asym():
     # Outputs a two section asymmetric wing surface
     # Set-up B-splines for each section. Done here since this information will be needed multiple times.
@@ -434,6 +448,48 @@ def get_two_section_surface_asym():
     }
 
     return surface, sec_chord_cp
+
+
+def get_single_section_surface():
+    """Create a dictionary with info and options about the aerodynamic
+    single section lifting surface"""
+
+    # Create a dictionary to store options about the mesh
+    mesh_dict = {
+        "num_y": 81,
+        "num_x": 2,
+        "wing_type": "rect",
+        "span": 2.0,
+        "root_chord": 1.0,
+        "symmetry": True,
+        "span_cos_spacing": 0,
+        "chord_cos_spacing": 0,
+    }
+
+    # Generate the aerodynamic mesh based on the previous dictionary
+    mesh = generate_mesh(mesh_dict)
+    surfaceSingle = {
+        # Wing definition
+        "name": "surface",  # name of the surface
+        "symmetry": True,  # if true, model one half of wing
+        # reflected across the plane y = 0
+        "S_ref_type": "wetted",  # how we compute the wing area,
+        # can be 'wetted' or 'projected'
+        "twist_cp": np.zeros(2),
+        "mesh": mesh,
+        "CL0": 0.0,  # CL of the surface at alpha=0
+        "CD0": 0.015,  # CD of the surface at alpha=0
+        # Airfoil properties for viscous drag calculation
+        "k_lam": 0.05,  # percentage of chord with laminar
+        # flow, used for viscous drag
+        "c_max_t": 0.303,  # chordwise location of maximum (NACA0015)
+        # thickness
+        "with_viscous": False,  # if true, compute viscous drag
+        "with_wave": False,  # if true, compute wave drag
+        "groundplane": False,
+    }
+
+    return surfaceSingle
 
 
 def assert_check_totals(totals, atol=1e-6, rtol=1e-6):
